[Dixit API Integration][Dixit Websocket Integration] Integrate Dixit server
